MERU entailment loss
Clamping the torch.asin() input to a maximum of 1.0 is necessary
for sensible half-aperture value. Otherwise, even modest entailment
weight (say 0.02) can push the text embedding into the epsilon-ball
in early training dynamics and render the output NaN. The official
MERU implementation actually clamps more aggresively but this is
not mentioned in the paper.